+++ MaxMind.GeoIP2/Model/Anonymizer.cs
@@ -1,9 +1,5 @@
-﻿#region
-
 using System;
 using System.Text.Json.Serialization;
-
-#endregion
 
 namespace MaxMind.GeoIP2.Model
 {
@@ -11,45 +7,8 @@ namespace MaxMind.GeoIP2.Model
     ///     Contains anonymizer-related data associated with an IP address.
     ///     This data is available from the GeoIP2 Insights web service.
     /// </summary>
-    public class Anonymizer
+    public record Anonymizer
     {
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public Anonymizer()
-        {
-        }
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public Anonymizer(
-            int? confidence = null,
-            bool isAnonymous = false,
-            bool isAnonymousVpn = false,
-            bool isHostingProvider = false,
-            bool isPublicProxy = false,
-            bool isResidentialProxy = false,
-            bool isTorExitNode = false,
-#if NET6_0_OR_GREATER
-            DateOnly? networkLastSeen = null,
-#endif
-            string? providerName = null
-        )
-        {
-            Confidence = confidence;
-            IsAnonymous = isAnonymous;
-            IsAnonymousVpn = isAnonymousVpn;
-            IsHostingProvider = isHostingProvider;
-            IsPublicProxy = isPublicProxy;
-            IsResidentialProxy = isResidentialProxy;
-            IsTorExitNode = isTorExitNode;
-#if NET6_0_OR_GREATER
-            NetworkLastSeen = networkLastSeen;
-#endif
-            ProviderName = providerName;
-        }
-
         /// <summary>
         ///     A score ranging from 1 to 99 that represents our percent confidence
         ///     that the network is currently part of an actively used VPN service.
@@ -57,7 +16,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("confidence")]
-        public int? Confidence { get; internal set; }
+        public int? Confidence { get; init; }
 
         /// <summary>
         ///     This is true if the IP address belongs to any sort of anonymous
@@ -65,7 +24,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("is_anonymous")]
-        public bool IsAnonymous { get; internal set; }
+        public bool IsAnonymous { get; init; }
 
         /// <summary>
         ///     This is true if the IP address is registered to an anonymous
@@ -79,7 +38,7 @@ namespace MaxMind.GeoIP2.Model
         /// </remarks>
         [JsonInclude]
         [JsonPropertyName("is_anonymous_vpn")]
-        public bool IsAnonymousVpn { get; internal set; }
+        public bool IsAnonymousVpn { get; init; }
 
         /// <summary>
         ///     This is true if the IP address belongs to a hosting or VPN
@@ -88,7 +47,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("is_hosting_provider")]
-        public bool IsHostingProvider { get; internal set; }
+        public bool IsHostingProvider { get; init; }
 
         /// <summary>
         ///     This is true if the IP address belongs to a public proxy.
@@ -96,7 +55,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("is_public_proxy")]
-        public bool IsPublicProxy { get; internal set; }
+        public bool IsPublicProxy { get; init; }
 
         /// <summary>
         ///     This is true if the IP address is on a suspected anonymizing
@@ -105,7 +64,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("is_residential_proxy")]
-        public bool IsResidentialProxy { get; internal set; }
+        public bool IsResidentialProxy { get; init; }
 
         /// <summary>
         ///     This is true if the IP address belongs to a Tor exit node.
@@ -113,7 +72,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("is_tor_exit_node")]
-        public bool IsTorExitNode { get; internal set; }
+        public bool IsTorExitNode { get; init; }
 
 #if NET6_0_OR_GREATER
         /// <summary>
@@ -123,7 +82,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("network_last_seen")]
-        public DateOnly? NetworkLastSeen { get; internal set; }
+        public DateOnly? NetworkLastSeen { get; init; }
 #endif
 
         /// <summary>
@@ -133,27 +92,6 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("provider_name")]
-        public string? ProviderName { get; internal set; }
-
-        /// <summary>
-        ///     Returns a <see cref="string" /> that represents this instance.
-        /// </summary>
-        /// <returns>
-        ///     A <see cref="string" /> that represents this instance.
-        /// </returns>
-        public override string ToString()
-        {
-            return $"{nameof(Confidence)}: {Confidence}, " +
-                $"{nameof(IsAnonymous)}: {IsAnonymous}, " +
-                $"{nameof(IsAnonymousVpn)}: {IsAnonymousVpn}, " +
-                $"{nameof(IsHostingProvider)}: {IsHostingProvider}, " +
-                $"{nameof(IsPublicProxy)}: {IsPublicProxy}, " +
-                $"{nameof(IsResidentialProxy)}: {IsResidentialProxy}, " +
-                $"{nameof(IsTorExitNode)}: {IsTorExitNode}, " +
-#if NET6_0_OR_GREATER
-                $"{nameof(NetworkLastSeen)}: {NetworkLastSeen}, " +
-#endif
-                $"{nameof(ProviderName)}: {ProviderName}";
-        }
+        public string? ProviderName { get; init; }
     }
 }