+++ MaxMind.GeoIP2/Model/RepresentedCountry.cs
@@ -1,48 +1,19 @@
-﻿#region
-
 using MaxMind.Db;
-using System.Collections.Generic;
 using System.Text.Json.Serialization;
-
-#endregion
 
 namespace MaxMind.GeoIP2.Model
 {
     /// <summary>
     ///     Contains data for the represented country associated with an IP address.
-    ///     This class contains the country-level data associated with an IP address for
+    ///     This record contains the country-level data associated with an IP address for
     ///     the IP's represented country. The represented country is the country
     ///     represented by something like a military base.
     ///     Do not use any of the country names as a database or dictionary
     ///     key. Use the <see cred="GeoNameId" /> or <see cred="IsoCode" />
     ///     instead.
     /// </summary>
-    public class RepresentedCountry : Country
+    public record RepresentedCountry : Country
     {
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public RepresentedCountry()
-        {
-        }
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        [Constructor]
-        public RepresentedCountry(
-            string? type = null,
-            int? confidence = null,
-            [Parameter("geoname_id")] long? geoNameId = null,
-            [Parameter("is_in_european_union")] bool isInEuropeanUnion = false,
-            [Parameter("iso_code")] string? isoCode = null,
-            IReadOnlyDictionary<string, string>? names = null,
-            IReadOnlyList<string>? locales = null)
-            : base(confidence, geoNameId, isInEuropeanUnion, isoCode, names, locales)
-        {
-            Type = type;
-        }
-
         /// <summary>
         ///     A string indicating the type of entity that is representing the
         ///     country. Currently we only return <c>military</c> but this could
@@ -50,6 +21,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("type")]
-        public string? Type { get; internal set; }
+        [MapKey("type")]
+        public string? Type { get; init; }
     }
 }