+++ MaxMind.GeoIP2/Model/Subdivision.cs
@@ -1,10 +1,5 @@
-﻿#region
-
 using MaxMind.Db;
-using System.Collections.Generic;
 using System.Text.Json.Serialization;
-
-#endregion
 
 namespace MaxMind.GeoIP2.Model
 {
@@ -14,31 +9,8 @@ namespace MaxMind.GeoIP2.Model
     ///     key. Use the <see cred="GeoNameId" /> or <see cred="IsoCode" />
     ///     instead.
     /// </summary>
-    public class Subdivision : NamedEntity
+    public record Subdivision : NamedEntity
     {
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public Subdivision()
-        {
-        }
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        [Constructor]
-        public Subdivision(
-            int? confidence = null,
-            [Parameter("geoname_id")] long? geoNameId = null,
-            [Parameter("iso_code")] string? isoCode = null,
-            IReadOnlyDictionary<string, string>? names = null,
-            IReadOnlyList<string>? locales = null)
-            : base(geoNameId, names, locales)
-        {
-            Confidence = confidence;
-            IsoCode = isoCode;
-        }
-
         /// <summary>
         ///     This is a value from 0-100 indicating MaxMind's confidence that
         ///     the subdivision is correct. This value is only set when using the
@@ -46,7 +18,8 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("confidence")]
-        public int? Confidence { get; internal set; }
+        [MapKey("confidence")]
+        public int? Confidence { get; init; }
 
         /// <summary>
         ///     This is a string up to three characters long contain the
@@ -59,6 +32,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("iso_code")]
-        public string? IsoCode { get; internal set; }
+        [MapKey("iso_code")]
+        public string? IsoCode { get; init; }
     }
 }