+++ MaxMind.GeoIP2/Responses/DomainResponse.cs
@@ -1,35 +1,13 @@
-﻿#region
-
 using MaxMind.Db;
 using System.Text.Json.Serialization;
-
-#endregion
 
 namespace MaxMind.GeoIP2.Responses
 {
     /// <summary>
-    ///     This class represents the GeoIP2 Domain response.
+    ///     This record represents the GeoIP2 Domain response.
     /// </summary>
-    /// <remarks>
-    /// Construct a DomainResponse model object.
-    /// </remarks>
-    /// <param name="domain"></param>
-    /// <param name="ipAddress"></param>
-    /// <param name="network"></param>
-    [method: Constructor]
-    public class DomainResponse(
-        string? domain,
-        [Inject("ip_address")] string? ipAddress,
-        [Network] Network? network = null
-        ) : AbstractResponse
+    public record DomainResponse : AbstractResponse
     {
-        /// <summary>
-        /// Construct a DomainResponse model object.
-        /// </summary>
-        public DomainResponse() : this(null, null)
-        {
-        }
-
         /// <summary>
         ///     The second level domain associated with the IP address. This will
         ///     be something like "example.com" or "example.co.uk", not
@@ -37,7 +15,8 @@ namespace MaxMind.GeoIP2.Responses
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("domain")]
-        public string? Domain { get; internal set; } = domain;
+        [MapKey("domain")]
+        public string? Domain { get; init; }
 
         /// <summary>
         ///     The IP address that the data in the model is for. If you
@@ -48,7 +27,8 @@ namespace MaxMind.GeoIP2.Responses
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("ip_address")]
-        public string? IPAddress { get; internal set; } = ipAddress;
+        [Inject("ip_address")]
+        public string? IPAddress { get; init; }
 
         /// <summary>
         ///     The network associated with the record. In particular, this is
@@ -57,6 +37,7 @@ namespace MaxMind.GeoIP2.Responses
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("network")]
-        public Network? Network { get; internal set; } = network;
+        [Network]
+        public Network? Network { get; init; }
     }
 }